ai_service/requirements.txt
fastapi
uvicorn
langchain-groq
langchain-core
python-dotenv
requests
pydantic
